Add runtime monitoring of CFO estimations for RF carrier control
An internal additional block is added to monitor the CFO
estimations and compare to a treshold. This block decides whether
or not to correct the CFO in software or in hardware.

To allow this modification, the NCO block was removed from the
coarse frequency recovery block. It should now be placed
externally. The same holds for the function probe block. The
reason is that the same function will need to adjust both the
NCO and the receiver hardare ocasionally, which would not be
possible if the NCO was internal to the block.