Add scheme for smoothing the NCO freq transitions
Also add exception to smoothed transition in the NCO.
Whenever the NCO is reset to 0 correction, the transition occurs
in a single step, rather than n_steps. This is because a specific
use case for this functionality is when the RTL center freq. is
adjusted and we want the NCO to change as fast as possible.