Add block to coordinate Rx logs
- Add an OOT module responsible for printing the receiver metrics
into console. The block is python-based and relies on threads that can
schedule the prints from each module with different periodicities. A
logger class is responsible for launching these threads.
- Move internal logs in the frame synchronization and SNR measurement
into this logger module.
- Add getters within frame synchronization and SNR measurement blocks
that allow the external logger to get the metrics of interest.
- Revert the verbosity parameter created in the MER measurement block,
as now the SNR prints are coordinated solely in the standalone logger
block.