Add split Rx flowgraphs with UDP communication
Introduce flowgraphs that implement distinct parts of the Rx chain,
the lower and the upper part, which UDP communication from one to the
other. The lower part is supposed to be executed in the host that is
attached to the RF interface, while the upper part can run on an
upstream host.